requirements.txt
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10
chromadb==0.4.22
openai>=1.10.0
python-dotenv==1.0.0
streamlit==1.29.0
pypdf2==3.0.1
python-docx==1.1.0
tiktoken==0.5.2